fix: Test if /etc/motd.d/insights-client points to an empty file
When `sos<=4.2` collects a report, `insights-client --offline` is
executed and updates the `motd` message on the file system. No files
should be changed while collecting sos report without using the
`--allow-system-change` option.

This patch adds a test to check if the "motd" message has been previously
disabled manually before registration. If
'/etc/motd.d/insights-client' already points to an empty file, the
`update_motd_message` function is ineffective.

Note that `sos>=4.2` doesn't execute `plugins:insights` anymore as it is
superfluous to collect twice the same data.

Resolves: rhbz#2008410

+++ src/insights_client/__init__.py
@@ -143,9 +143,14 @@ def update_motd_message():
     It is intentional that the message does not reappear if a system is then
     unregistered. Only if both the unregistered and the registered stamp files
     do not exist is an motd symlink created.
+
+    The motd message could also be deliberately disabled by the users before
+    registration, simply because they don't want to use insights-client, by
+    pointing /etc/motd.d/insights-client at an empty file.
     """
     try:
-        if os.path.exists(os.path.dirname(MOTD_FILE)):
+        if os.path.exists(os.path.dirname(MOTD_FILE)) and \
+           os.path.islink(MOTD_FILE) and not os.path.samefile(os.devnull, MOTD_FILE):
             if (os.path.isfile(REGISTERED_FILE) or os.path.isfile(UNREGISTERED_FILE)):
                 if not os.path.samefile(os.devnull, MOTD_FILE):
                     os.symlink(os.devnull, MOTD_FILE + ".tmp")
